Implemented function to add, fetch and display from the API
